Feat/jwt

+++ x/association/handler.go
@@ -2,10 +2,12 @@
 package association
 
 import (
-    "errors"
-    "net/http"
+	"errors"
+	"net/http"
+
+	"github.com/labstack/echo/v4"
+	"github.com/totegamma/concurrent/x/util"
 	"gorm.io/gorm"
-    "github.com/labstack/echo/v4"
 )
 
 // Handler handles Association objects
@@ -51,11 +53,21 @@ func (h Handler) Post(c echo.Context) error {
 
 // Delete is for Handling HTTP Delete Method
 func (h Handler) Delete(c echo.Context) error {
-
     var request deleteQuery
     err := c.Bind(&request)
     if (err != nil) {
         return err
+    }
+
+    target, err := h.service.Get(request.ID)
+    if err != nil {
+        return c.JSON(http.StatusNotFound, echo.Map{"error": "target association not found"})
+    }
+
+    claims := c.Get("jwtclaims").(util.JwtClaims)
+    requester := claims.Audience
+    if target.Author != requester {
+        return c.JSON(http.StatusForbidden, echo.Map{"error": "you are not authorized to perform this action"})
     }
 
     err = h.service.Delete(request.ID)
